Updated Access Control

+++ Clean.Application/Services/InventoryCommentService/InventoryCommentService.cs
@@ -1,13 +1,20 @@
+using System.Security.Claims;
 using Clean.Application.Abstractions;
 using Clean.Application.Dtos;
 using Clean.Application.Filters;
 using Clean.Application.Responses;
 using Clean.Domain.Entities;
+using Clean.Domain.Entities.Enums;
+using Microsoft.AspNetCore.Http;
 
 namespace Clean.Application.Services;
 
-public class InventoryCommentService(IInventoryCommentRepository repository) : IInventoryCommentService
+public class InventoryCommentService(IInventoryCommentRepository repository
+    ,IHttpContextAccessor httpContextAccessor,IUserRepository userRepository, IInventoryCommentNotifier notifier) : IInventoryCommentService
 {
+    private int? GetCurrentUserId() =>
+        int.TryParse(httpContextAccessor.HttpContext?.User?.FindFirst(ClaimTypes.NameIdentifier)
+            ?.Value, out var userId) ? userId : null;
     public async Task<PagedResponse<InventoryCommentGetDto>> GetAll(InventoryCommentFilter filter)
     {
         var result = await repository.GetAll(filter);
@@ -39,20 +46,43 @@ public class InventoryCommentService(IInventoryCommentRepository repository) : I
 
     public async Task<Response<string>> Create(InventoryCommentCreateDto dto)
     {
+        var curentUserId = GetCurrentUserId();
+        if(curentUserId == null)
+            return new Response<string>(400,"Not Authorized");
         var model = new InventoryComment
         {
             InventoryId = dto.InventoryId,
             UserId = dto.UserId,
-            Content = dto.Content
+            Content = dto.Content,
+            CreatedAt = DateTime.UtcNow
         };
+        var commentDto = new InventoryCommentGetDto
+        {
+            InventoryId = model.InventoryId,
+            UserId = model.UserId,
+            Content = model.Content,
+            CreatedAt = model.CreatedAt
+        };
+
+        await notifier.NotifyCommentCreated(dto.InventoryId, commentDto);
         await repository.Create(model);
         return new Response<string>(200, "Inventory comment created");
     }
 
     public async Task<Response<string>> Delete(int id)
     {
+        var curentUserId = GetCurrentUserId();
+        if(curentUserId == null)
+            return new Response<string>(403,"Not Authorized");
+        var comment = await repository.GetById(id);
+        var usr = await userRepository.GetById((int)curentUserId);
+        if (comment.UserId != curentUserId && usr.Role!=UserRole.Admin)
+        {
+            return new Response<string>(403,"Not Authorized");
+        }
         var result = await repository.Delete(id);
         if (result == -1) return new Response<string>(404, "Inventory comment not found");
+        await notifier.NotifyCommentDeleted(comment.InventoryId, id);
         return new Response<string>(200, "Inventory comment deleted");
     }
 }
